ストアをモジュールに分割
