+++ resources/js/mypage/app.js
@@ -3,7 +3,7 @@ import './plugins';
 import App from './App.vue';
 
 import store from './store';
-import router from './router'
+import router from './plugins/router'
 
 // components
 const files = require.context('./Components', true, /\.vue$/i);
