+++ resources/js/mypage/store/index.js
@@ -1,25 +1,27 @@
 import Vue from 'vue';
 import Vuex from 'vuex'
-import api from '../api'
-import { DateTime } from "luxon";
+
+import ArticlesModule from './modules/articles';
+import AuthModule from './modules/auth';
+import AttachmentsModule from './modules/attachments';
+import TagsModule from './modules/tags';
+import AnalyticsModule from './modules/analytics';
+import OptionsModule from './modules/options';
 Vue.use(Vuex);
 
-const SET_INITIALIZED = 'SET_INITIALIZED';
 const SET_INFO_MESSAGE = 'SET_INFO_MESSAGE';
 const SET_API_STATUS = 'SET_API_STATUS';
-const SET_USER = 'SET_USER';
-const SET_ARTICLES = 'SET_ARTICLES';
-const SET_OPTIONS = 'SET_OPTIONS';
-const SET_ATTACHMENTS = 'SET_ATTACHMENTS';
-const SET_TAGS = 'SET_TAGS';
-const SET_ANALYTICS = 'SET_ANALYTICS';
 
 export default new Vuex.Store({
+  modules: {
+    AuthModule,
+    ArticlesModule,
+    AttachmentsModule,
+    TagsModule,
+    OptionsModule,
+    AnalyticsModule,
+  },
   state: {
-    /**
-     * 初期化ステータス（ログイン判定完了で初期化完了）
-     */
-    initialized: false,
     /**
      * 通知メッセージ
      */
@@ -33,18 +35,8 @@ export default new Vuex.Store({
       message: null,
       status_code: null,
     },
-    /**
-     * 各種データ。false:未ロード
-     */
-    user: false,
-    articles: false,
-    options: false,
-    attachments: false,
-    tags: false,
-    analytics: [],
   },
   getters: {
-    initialized: state => state.initialized,
     infoMessage: state => state.info_message,
     // api
     fetching: state => state.api_status.fetching,
@@ -63,32 +55,8 @@ export default new Vuex.Store({
       }
       return null;
     },
-    // user
-    user: state => state.user,
-    isLoggedIn: state => !!state.user,
-    isAdmin: state => state.user && state.user.admin,
-    isVerified: state => state.user && state.user.verified,
-
-    // article
-    articlesLoaded: state => state.articles !== false,
-    articles: state => state.articles || [],
-    // option
-    optionsLoaded: state => state.options !== false,
-    options: state => state.options || {},
-    // attachment
-    attachmentsLoaded: state => state.attachments !== false,
-    attachments: state => state.attachments || [],
-    // tag
-    tagsLoaded: state => state.tags !== false,
-    tags: state => state.tags || [],
-    // analytic
-    analyticsLoaded: state => state.analytics !== false,
-    analytics: state => state.analytics || [],
   },
   mutations: {
-    [SET_INITIALIZED](state, initialized = false) {
-      state.initialized = initialized
-    },
     [SET_INFO_MESSAGE](state, message = null) {
       state.info_message = message;
     },
@@ -97,30 +65,6 @@ export default new Vuex.Store({
       state.api_status.errors = errors;
       state.api_status.message = message;
       state.api_status.status_code = status_code;
-    },
-    [SET_USER](state, user = false) {
-      state.user = user;
-    },
-    [SET_ARTICLES](state, articles = false) {
-      if (articles) {
-        articles = articles.map(a => Object.assign(a, {
-          created_at: DateTime.fromISO(a.created_at),
-          updated_at: DateTime.fromISO(a.updated_at),
-        }));
-      }
-      state.articles = articles;
-    },
-    [SET_OPTIONS](state, options = false) {
-      state.options = options;
-    },
-    [SET_ATTACHMENTS](state, attachments = false) {
-      state.attachments = attachments;
-    },
-    [SET_TAGS](state, tags = false) {
-      state.tags = tags;
-    },
-    [SET_ANALYTICS](state, analytics = false) {
-      state.analytics = analytics;
     },
   },
   actions: {
@@ -170,277 +114,5 @@ export default new Vuex.Store({
       }
       return commit(SET_API_STATUS, { message: 'エラーが発生しました', status_code: res.status });
     },
-
-    // 認証系
-    /**
-     * ユーザーを取得できるかでログイン済みかを判定する。
-     * 判定後に初期化ステータスを完了にする
-     */
-    async checkLogin({ commit }) {
-      try {
-        const res = await api.fetchUser();
-        if (res && res.status === 200) {
-          commit(SET_USER, res.data.data);
-        }
-      } catch (e) { }
-      commit(SET_INITIALIZED, true);
-    },
-    /**
-     * ログイン
-     */
-    async login({ commit, dispatch }, params) {
-      dispatch('setApiStatusFetching');
-      try {
-        const res = await api.login(params);
-        if (res && res.status === 200) {
-          dispatch('setApiStatusSuccess');
-          return commit(SET_USER, res.data.data);
-        }
-        dispatch('setApiStatusError');
-      } catch (e) {
-        dispatch('setApiStatusError', e);
-      }
-    },
-    async logout({ commit }) {
-      try {
-        api.logout();
-        commit(SET_USER, null);
-      } catch (e) { }
-    },
-    async register({ commit, dispatch }, params) {
-      dispatch('setApiStatusFetching');
-      try {
-        const res = await api.register(params);
-        if (res && res.status === 201) {
-          commit(SET_USER, res.data.data);
-          dispatch('setApiStatusSuccess');
-          return dispatch('setInfoMessage', { message: 'ユーザー登録しました。' });
-        }
-        dispatch('setApiStatusError');
-      } catch (e) {
-        dispatch('setApiStatusError', e);
-      }
-    },
-    async sendResetEmail({ dispatch }, params) {
-      dispatch('setApiStatusFetching');
-      try {
-        const res = await api.reset(params);
-        if (res && res.status === 200) {
-          dispatch('setApiStatusSuccess');
-          return dispatch('setInfoMessage', { message: 'メールを送信しました。' });
-        }
-        dispatch('setApiStatusError');
-      } catch (e) {
-        dispatch('setApiStatusError', e);
-      }
-    },
-    /**
-     * 確認メール送信
-     */
-    async sendVerifyEmail({ dispatch }) {
-      dispatch('setApiStatusFetching');
-      try {
-        const res = await api.resend();
-        if (res && res.status === 200) {
-          dispatch('setApiStatusSuccess');
-          return dispatch('setInfoMessage', { message: 'メールを送信しました。' });
-        }
-        dispatch('setApiStatusError');
-      } catch (e) {
-        dispatch('setApiStatusError', e);
-      }
-    },
-
-    /**
-     * 記事編集用オプション
-     */
-    async fetchOptions({ commit }) {
-      try {
-        const res = await api.fetchOptions();
-        if (res && res.status === 200) {
-          return commit(SET_OPTIONS, res.data);
-        }
-        dispatch('setApiStatusError');
-      } catch (e) {
-        dispatch('setApiStatusError', e);
-      }
-    },
-
-    // 記事
-    /**
-     * 投稿記事一覧
-     */
-    async fetchArticles({ commit }) {
-      try {
-        const res = await api.fetchArticles();
-        if (res && res.status === 200) {
-          return commit(SET_ARTICLES, res.data.data);
-        }
-        dispatch('setApiStatusError');
-      } catch (e) {
-        dispatch('setApiStatusError', e);
-      }
-    },
-    /**
-     * 記事更新
-     */
-    async updateArticle({ commit, dispatch }, { params, message = '更新しました' }) {
-      dispatch('setApiStatusFetching');
-      try {
-        const res = await api.updateArticle(params);
-        if (res && res.status === 200) {
-          dispatch('setApiStatusSuccess');
-          if (params.preview) {
-            return res.data;
-          }
-          commit(SET_ARTICLES, res.data.data);
-          return dispatch('setInfoMessage', { message });
-        }
-        dispatch('setApiStatusError');
-      } catch (e) {
-        dispatch('setApiStatusError', e);
-      }
-    },
-    /**
-     * 記事作成
-     */
-    async createArticle({ commit, dispatch }, { params, message = '作成しました' }) {
-      dispatch('setApiStatusFetching');
-      try {
-        const res = await api.createArticle(params);
-        if (res && res.status === 200) {
-          dispatch('setApiStatusSuccess');
-          if (params.preview) {
-            return res.data;
-          }
-          commit(SET_ARTICLES, res.data.data);
-          return dispatch('setInfoMessage', { message });
-        }
-        dispatch('setApiStatusError');
-      } catch (e) {
-        dispatch('setApiStatusError', e);
-      }
-    },
-
-    // ファイル
-    /**
-     * ファイル一覧
-     */
-    async fetchAttachments({ commit }) {
-      try {
-        const res = await api.fetchAttachments();
-        if (res && res.status === 200) {
-          return commit(SET_ATTACHMENTS, res.data.data);
-        }
-        dispatch('setApiStatusError');
-      } catch (e) {
-        dispatch('setApiStatusError', e);
-      }
-    },
-    /**
-     * ファイル投稿
-     */
-    async storeAttachment({ commit, dispatch }, { file, type, id = null, only_image = false }) {
-      dispatch('setApiStatusFetching');
-      try {
-        const formData = new FormData();
-        formData.append("file", file);
-        formData.append("type", type);
-        if (id) {
-          formData.append("id", id);
-        }
-        formData.append("only_image", only_image ? 1 : 0);
-
-        const res = await api.storeAttachment(formData);
-        if (res && res.status === 200) {
-          commit(SET_ATTACHMENTS, res.data.data);
-          return dispatch('setApiStatusSuccess');
-        }
-        dispatch('setApiStatusError');
-      } catch (e) {
-        dispatch('setApiStatusError', e);
-      }
-    },
-    /**
-     * ファイル削除
-     */
-    async deleteAttachment({ commit, dispatch }, id) {
-      dispatch('setApiStatusFetching');
-      try {
-        const res = await api.deleteAttachment(id);
-        if (res && res.status === 200) {
-          commit(SET_ATTACHMENTS, res.data.data);
-          return dispatch('setApiStatusSuccess');
-        }
-        dispatch('setApiStatusError');
-      } catch (e) {
-        dispatch('setApiStatusError', e);
-      }
-    },
-
-    // タグ
-    /**
-     * タグ検索
-     */
-    async fetchTags({ commit, dispatch }, name = '') {
-      try {
-        const res = await api.fetchTags(name);
-        if (res && res.status === 200) {
-          return commit(SET_TAGS, res.data);
-        }
-        dispatch('setApiStatusError');
-      } catch (e) {
-        dispatch('setApiStatusError', e);
-      }
-    },
-    /**
-     * タグ登録
-     */
-    async storeTag({ dispatch }, name) {
-      try {
-        const res = await api.storeTag(name);
-        if (res && res.status === 201) {
-          return;
-        }
-        dispatch('setApiStatusError');
-      } catch (e) {
-        dispatch('setApiStatusError', e);
-      }
-    },
-
-    /**
-     * ユーザー更新
-     */
-    async updateUser({ commit, dispatch }, { user, message = 'プロフィールを更新しました。' }) {
-      dispatch('setApiStatusFetching');
-      try {
-        const res = await api.updateUser(user);
-        if (res && res.status === 200) {
-          commit(SET_USER, res.data.data);
-          dispatch('setApiStatusSuccess');
-          return dispatch('setInfoMessage', { message });
-        }
-        dispatch('setApiStatusError');
-      } catch (e) {
-        dispatch('setApiStatusError', e);
-      }
-    },
-    /**
-     * 分析データ取得
-     */
-    async fetchAnalytics({ commit, dispatch }, params) {
-      dispatch('setApiStatusFetching');
-      try {
-        const res = await api.fetchAnalytics(params)
-        if (res && res.status === 200) {
-          commit(SET_ANALYTICS, res.data.data);
-          return dispatch('setApiStatusSuccess');
-        }
-        dispatch('setApiStatusError');
-      } catch (e) {
-        dispatch('setApiStatusError', e);
-      }
-    }
-
   }
-})
+});
